ENH Implement add_env option for Job and grid_map
The add_env option allows to specify additional environment variables
for a new instance of job, or all jobs by grid_map.
It is a dictionary that contains key, value pairs where key is the name
of the environment variable.
Each environment variable in add_env that is already present in the job
environment due to a copied environment (through ``copy_env=True``), is
overwritten by the value in add_env.